@@ -22,8 +22,8 @@
     }
   },
   "require": {
-    "php": "^7.2",
-    "guzzlehttp/guzzle": "^6.3",
+    "php": "^7.4",
+    "guzzlehttp/guzzle": "^6.5",
     "monolog/monolog": "^2.0",
     "ext-json": "*",
     "ext-libxml": "*",
@@ -32,8 +32,8 @@
   "require-dev": {
     "roave/security-advisories": "dev-master",
     "phpunit/phpunit": "^9.0",
-    "squizlabs/php_codesniffer": "^3",
-    "php-coveralls/php-coveralls": "^2"
+    "squizlabs/php_codesniffer": "^3.5",
+    "php-coveralls/php-coveralls": "^2.2"
   },
   "scripts": {
     "test": [
